server/requirements.txt
flask==2.0.1
flask-socketio==5.3.0
eventlet==0.33.3
python-dotenv==0.19.0